+++ R/load_yaml_files.R
@@ -49,7 +49,7 @@ load_yaml_files <- function() {
       fixed = TRUE
     )
 
-  yamls_default <- lapply(X = yaml_files, FUN = yaml::read_yaml)
+  yamls_default <- furrr::future_map(.x = yaml_files, .f = yaml::read_yaml)
 
   names(yamls_default) <- yaml_names